+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2022-11-04
+
+### Added
+- added class org.adwmainz.da.extensions.askmore.operations.FullySelectElementsOperation.java
+
 ## [1.3.0] - 2020-07-30
 
 ### Added